Move DataHash tests to their own file
Now that DataHash lives in picard.util.datahash, move its unit tests out
of test/test_coverart_image.py into test/test_util_datahash.py. The
CoverArtImage ordering test that happens to exercise datahash ordering
(test_lt_datahash) stays with the image tests.
